+++ pkg/repositories/repositories.go
@@ -35,18 +35,40 @@ type ListResult struct {
 	Repositories  []Repo
 }
 
-// GithubRepo is a sparce struct for just the GitHub repository info we need.
-type GithubRepo struct {
-	Name        string
+// A sparce struct for caching just the GitHub
+// repository details we need.
+type GithubRepoDetails struct {
+	// Repository name
+	Name string
+
+	// Repository description
 	Description string
-	IsPrivate   bool
+
+	// Name of the default branch
+	DefaultBranch string
+
+	// Whether the repository is private. If false, it's public.
+	IsPrivate bool
+}
+
+// A struct for caching repository content information.
+type GithubRepoContentDetails struct {
+	// Whether the repository has a CircleCI configuration file.
+	HasCircleCI bool
+
+	// Whether the repository has a README.md in the root directory.
+	HasReadme bool
 }
 
 type Service struct {
 	config       Config
 	ctx          context.Context
 	githubClient *github.Client
-	githubRepos  map[string]GithubRepo
+
+	// Cached information on certain repos
+	githubRepoDetails map[string]GithubRepoDetails
+	// Cached information on repo content
+	githubRepoContentDetails map[string]GithubRepoContentDetails
 }
 
 // New instantiates a new repositories service.
@@ -70,41 +92,46 @@ func New(c Config) (*Service, error) {
 	client := github.NewClient(tc)
 
 	s := &Service{
-		config:       c,
-		ctx:          ctx,
-		githubClient: client,
+		config:                   c,
+		ctx:                      ctx,
+		githubClient:             client,
+		githubRepoDetails:        make(map[string]GithubRepoDetails),
+		githubRepoContentDetails: make(map[string]GithubRepoContentDetails),
 	}
 
-	repos, err := s.loadGithubRepoData()
+	err := s.loadGithubRepoDetails()
 	if err != nil {
-		return nil, microerror.Mask(err)
+		return nil, err
 	}
-
-	s.githubRepos = repos
 
 	return s, nil
 }
 
-// Load repository metadata from Github.
-func (s *Service) loadGithubRepoData() (map[string]GithubRepo, error) {
+// Load main information for all repositories of the organization.
+//
+// This also loads info for archived repos, but it's still cheaper
+// to do it this way than one by one, as we get up to 100 repos
+// per request.
+func (s *Service) loadGithubRepoDetails() error {
 	opts := &github.RepositoryListByOrgOptions{
 		ListOptions: github.ListOptions{
 			PerPage: 100,
 		},
 	}
-	repos := make(map[string]GithubRepo)
+	repos := make(map[string]GithubRepoDetails)
 
 	for {
 		r, resp, err := s.githubClient.Repositories.ListByOrg(s.ctx, s.config.GithubOrganization, opts)
 		if err != nil {
-			return nil, err
+			return err
 		}
 
 		for _, repo := range r {
-			repos[repo.GetName()] = GithubRepo{
-				Name:        repo.GetName(),
-				Description: repo.GetDescription(),
-				IsPrivate:   repo.GetPrivate(),
+			repos[repo.GetName()] = GithubRepoDetails{
+				Name:          repo.GetName(),
+				Description:   repo.GetDescription(),
+				IsPrivate:     repo.GetPrivate(),
+				DefaultBranch: repo.GetDefaultBranch(),
 			}
 		}
 
@@ -114,7 +141,33 @@ func (s *Service) loadGithubRepoData() (map[string]GithubRepo, error) {
 		opts.Page = resp.NextPage
 	}
 
-	return repos, nil
+	s.githubRepoDetails = repos
+
+	return nil
+}
+
+func (s *Service) loadGithubRepoContentDetails(name string) error {
+	details := GithubRepoContentDetails{}
+
+	_, _, resp, err := s.githubClient.Repositories.GetContents(s.ctx, s.config.GithubOrganization, name, ".circleci/config.yml", nil)
+	if err == nil {
+		details.HasCircleCI = true
+	} else if resp.StatusCode != 404 {
+		// 404 is a "not found" error, which is expected. Everything else is not expected.
+		return err
+	}
+
+	_, _, resp, err = s.githubClient.Repositories.GetContents(s.ctx, s.config.GithubOrganization, name, "README.md", nil)
+	if err == nil {
+		details.HasReadme = true
+	} else if resp.StatusCode != 404 {
+		// 404 is a "not found" error, which is expected. Everything else is not expected.
+		return err
+	}
+
+	s.githubRepoContentDetails[name] = details
+
+	return nil
 }
 
 // Loads a list of repository configurations from a local path.
@@ -176,19 +229,53 @@ func (s *Service) GetLists() ([]ListResult, error) {
 }
 
 // Returns the description for the given repo. If not available,
-// returns an empty string.
-func (s *Service) GetDescription(name string) string {
-	if repo, ok := s.githubRepos[name]; ok {
-		return repo.Description
+// or an error occurs, returns an empty string.
+func (s *Service) MustGetDescription(name string) string {
+	if _, ok := s.githubRepoDetails[name]; !ok {
+		return ""
 	}
-	return ""
+
+	return s.githubRepoDetails[name].Description
 }
 
-// Returns the public/private info for the given repo. If not available,
-// return an error.
+// Returns the public/private info for the given repo.
 func (s *Service) GetIsPrivate(name string) (bool, error) {
-	if repo, ok := s.githubRepos[name]; ok {
-		return repo.IsPrivate, nil
+	if _, ok := s.githubRepoDetails[name]; !ok {
+		return false, microerror.Maskf(repositoryNotFoundError, "repository %s not found", name)
 	}
-	return false, microerror.Maskf(repositoryNotFoundError, "repository %s not found", name)
+
+	return s.githubRepoDetails[name].IsPrivate, nil
+}
+
+// Returns the default branch name. Returns an empty string in case of error.
+func (s *Service) MustGetDefaultBranch(name string) string {
+	if _, ok := s.githubRepoDetails[name]; !ok {
+		return ""
+	}
+
+	return s.githubRepoDetails[name].DefaultBranch
+}
+
+// Returns whether the repo has a CircleCI configuration.
+func (s *Service) GetHasCircleCI(name string) (bool, error) {
+	if _, ok := s.githubRepoContentDetails[name]; !ok {
+		err := s.loadGithubRepoContentDetails(name)
+		if err != nil {
+			return false, microerror.Mask(err)
+		}
+	}
+
+	return s.githubRepoContentDetails[name].HasCircleCI, nil
+}
+
+// Returns whether the repo has a main README file.
+func (s *Service) GetHasReadme(name string) (bool, error) {
+	if _, ok := s.githubRepoContentDetails[name]; !ok {
+		err := s.loadGithubRepoContentDetails(name)
+		if err != nil {
+			return false, microerror.Mask(err)
+		}
+	}
+
+	return s.githubRepoContentDetails[name].HasReadme, nil
 }